problemoSju

+++ src/Input/ReadFromFile.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 public class ReadFromFile{
     private final List<String> elements = new ArrayList<>();
-    private final List<Long> numbers = new ArrayList<>();
+    private final List<Integer> numbers = new ArrayList<>();
 
     private final List<List<String>> groups = new ArrayList<>();
     BufferedReader br;
@@ -39,7 +39,7 @@ public class ReadFromFile{
         return elements;
     }
 
-    public List<Long> seperatedByComma(File fileName){
+    public List<Integer> seperatedByComma(File fileName){
         try{
             String line = null;
             br = new BufferedReader(new FileReader(fileName));
@@ -48,7 +48,7 @@ public class ReadFromFile{
                 String[] values = line.split(",");
                 for (String str : values) {
                     //elements.add(str);
-                    numbers.add(Long.parseLong(str));
+                    numbers.add(Integer.parseInt(str));
                 }
             }
             br.close();
